requirements.txt
Flask
flask_cors
gunicorn
requests
python-dotenv
numpy